+++ tidy3d/__init__.py
@@ -186,9 +186,14 @@ from .components.medium import (
     TwoPhotonAbsorption,
     medium_from_nk,
 )
+from .components.mode.data.sim_data import ModeSimulationData
+from .components.mode.run_local import run_local
+
+# Mode simulation
+from .components.mode.simulation import ModeSimulation
 
 # modes
-from .components.mode import ModeSpec
+from .components.mode_spec import ModeSpec
 
 # monitors
 from .components.monitor import (
@@ -534,4 +539,7 @@ __all__ = [
     "EMELengthSweep",
     "EMEModeSweep",
     "EMEFreqSweep",
+    "ModeSimulation",
+    "ModeSimulationData",
+    "run_local",
 ]
