+++ tidy3d/components/mode_spec.py
