+++ tidy3d/components/monitor.py
@@ -13,7 +13,7 @@ from .apodization import ApodizationSpec
 from .base import Tidy3dBaseModel, cached_property, skip_if_fields_missing
 from .base_sim.monitor import AbstractMonitor
 from .medium import MediumType
-from .mode import ModeSpec
+from .mode_spec import ModeSpec
 from .types import (
     ArrayFloat1D,
     Ax,
