+++ tidy3d/web/api/tidy3d_stub.py
@@ -17,6 +17,8 @@ from ...components.eme.simulation import EMESimulation
 from ...components.heat_charge.heat.simulation import HeatSimulation
 from ...components.heat_charge.sim_data import HeatChargeSimulationData, HeatSimulationData
 from ...components.heat_charge.simulation import HeatChargeSimulation
+from ...components.mode.data.sim_data import ModeSimulationData
+from ...components.mode.simulation import ModeSimulation
 from ...components.simulation import Simulation
 from ...plugins.mode.mode_solver import ModeSolver
 from ..core.file_util import (
@@ -27,9 +29,15 @@ from ..core.file_util import (
 from ..core.stub import TaskStub, TaskStubData
 from ..core.types import TaskType
 
-SimulationType = Union[Simulation, HeatChargeSimulation, HeatSimulation, EMESimulation]
+SimulationType = Union[
+    Simulation, HeatChargeSimulation, HeatSimulation, EMESimulation, ModeSimulation
+]
 SimulationDataType = Union[
-    SimulationData, HeatChargeSimulationData, HeatSimulationData, EMESimulationData
+    SimulationData,
+    HeatChargeSimulationData,
+    HeatSimulationData,
+    EMESimulationData,
+    ModeSimulationData,
 ]
 
 
@@ -45,11 +53,11 @@ class Tidy3dStub(BaseModel, TaskStub):
         ----------
         file_path : str
             Full path to the .yaml or .json or .hdf5 file to load the
-            Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`] from.
+            Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`, :class:`.ModeSimulation`] from.
 
         Returns
         -------
-        Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`]
+        Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`, :class:`.ModeSimulation`]
             An instance of the component class calling ``load``.
 
         Example
@@ -76,6 +84,8 @@ class Tidy3dStub(BaseModel, TaskStub):
             sim = HeatChargeSimulation.from_file(file_path)
         elif "EMESimulation" == type_:
             sim = EMESimulation.from_file(file_path)
+        elif "ModeSimulation" == type_:
+            sim = ModeSimulation.from_file(file_path)
 
         return sim
 
@@ -83,7 +93,7 @@ class Tidy3dStub(BaseModel, TaskStub):
         self,
         file_path: str,
     ):
-        """Exports Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`] instance to .yaml, .json,
+        """Exports Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`, :class:`.ModeSimulation`] instance to .yaml, .json,
         or .hdf5 file
 
         Parameters
@@ -98,7 +108,7 @@ class Tidy3dStub(BaseModel, TaskStub):
         self.simulation.to_file(file_path)
 
     def to_hdf5_gz(self, fname: str, custom_encoders: List[Callable] = None) -> None:
-        """Exports Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`] instance to .hdf5.gz file.
+        """Exports Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`, :class:`.ModeSimulation`] instance to .hdf5.gz file.
 
         Parameters
         ----------
@@ -134,6 +144,8 @@ class Tidy3dStub(BaseModel, TaskStub):
             return TaskType.HEAT.name
         elif isinstance(self.simulation, EMESimulation):
             return TaskType.EME.name
+        elif isinstance(self.simulation, ModeSimulation):
+            return TaskType.MODE.name
 
     def validate_pre_upload(self, source_required) -> None:
         """Perform some pre-checks on instances of component"""
@@ -150,18 +162,18 @@ class Tidy3dStubData(BaseModel, TaskStubData):
 
     @classmethod
     def from_file(cls, file_path: str) -> SimulationDataType:
-        """Loads a Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`]
+        """Loads a Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`, :class:`.ModeSimulationData`]
         from .yaml, .json, or .hdf5 file.
 
         Parameters
         ----------
         file_path : str
             Full path to the .yaml or .json or .hdf5 file to load the
-            Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`] from.
+            Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`, :class:`.ModeSimulationData`] from.
 
         Returns
         -------
-        Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`]
+        Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`, :class:`.ModeSimulationData`]
             An instance of the component class calling ``load``.
         """
         extension = _get_valid_extension(file_path)
@@ -184,18 +196,20 @@ class Tidy3dStubData(BaseModel, TaskStubData):
             sim_data = HeatChargeSimulationData.from_file(file_path)
         elif "EMESimulationData" == type_:
             sim_data = EMESimulationData.from_file(file_path)
+        elif "ModeSimulationData" == type_:
+            sim_data = ModeSimulationData.from_file(file_path)
 
         return sim_data
 
     def to_file(self, file_path: str):
-        """Exports Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`] instance
+        """Exports Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`, :class:`.ModeSimulationData`] instance
         to .yaml, .json, or .hdf5 file
 
         Parameters
         ----------
         file_path : str
             Full path to the .yaml or .json or .hdf5 file to save the
-            Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`] to.
+            Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`, :class:`.ModeSimulationData`] to.
 
         Example
         -------
@@ -206,17 +220,17 @@ class Tidy3dStubData(BaseModel, TaskStubData):
     @classmethod
     def postprocess(cls, file_path: str) -> SimulationDataType:
         """Load .yaml, .json, or .hdf5 file to
-        Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`] instance.
+        Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`, :class:`.ModeSimulationData`] instance.
 
         Parameters
         ----------
         file_path : str
             Full path to the .yaml or .json or .hdf5 file to save the
-            Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`] to.
+            Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`, :class:`.ModeSimulationData`] to.
 
         Returns
         -------
-        Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`]
+        Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`, :class:`.ModeSimulationData`]
             An instance of the component class calling ``load``.
         """
         stub_data = Tidy3dStubData.from_file(file_path)
